Only export the selected armature
Require that the user select an armature for export. Attempt to export
all actions as if they belong to that armature. This is not a great
way to do things, but right now it's the easiest way.

In Blender's internal model, actions are a set of channels with each
channel being associated with exactly one bone, and a given bone being
associated with exactly one armature. Therefore, an Action could
hypothetically be considered to "belong" to more than one armature
at a time.

A future update should, for each Action, check that the action has at
least one channel associated with a bone in the selected armature.
If there is no association, the Action won't be exported for the
armature.